Authentication added  (#82)
* WIP

* Authentication backend looks fine

* WIP

* WIP

* Saving user properly

* Authentication OK

* Login page and some other stuffs

---------

+++ canopeum_backend/canopeum_backend/management/commands/initialize_database.py
@@ -18,11 +18,13 @@ from canopeum_backend.models import (
     Mulchlayertype,
     MulchlayertypeInternationalization,
     Post,
+    Role,
     Site,
     Sitetype,
     SitetypeInternationalization,
     TreespeciestypeInternationalization,
     Treetype,
+    User,
 )
 
 
@@ -37,22 +39,24 @@ class Command(BaseCommand):
         if response != "yes":
             self.stdout.write(self.style.ERROR("Operation cancelled"))
             return
-        self.stdout.write("Erasing existing data...")
-        assets_to_delete = Asset.objects.all().exclude(asset="site_img.png")
-        for asset in assets_to_delete:
-            path = Path(settings.BASE_DIR) / "canopeum_backend" / "media" / asset.asset.name
-            path.unlink(missing_ok=True)
-        call_command("flush", "--noinput")
         with connection.cursor() as cursor:
-            cursor.execute("SET FOREIGN_KEY_CHECKS = 0;")
-            cursor.execute("SHOW TABLES;")
-            tables = cursor.fetchall()
-            for table in tables:
-                cursor.execute(f"DROP TABLE IF EXISTS {table[0]};")
-            cursor.execute("SET FOREIGN_KEY_CHECKS = 1;")
+            if cursor.execute("SHOW TABLES;") != 0:
+                self.stdout.write("Erasing existing data...")
+                assets_to_delete = Asset.objects.all().exclude(asset="site_img.png")
+                for asset in assets_to_delete:
+                    path = Path(settings.BASE_DIR) / "canopeum_backend" / "media" / asset.asset.name
+                    path.unlink(missing_ok=True)
+                call_command("flush", "--noinput")
+                cursor.execute("SET FOREIGN_KEY_CHECKS = 0;")
+                cursor.execute("SHOW TABLES;")
+                tables = cursor.fetchall()
+                for table in tables:
+                    cursor.execute(f"DROP TABLE IF EXISTS {table[0]};")
+                cursor.execute("SET FOREIGN_KEY_CHECKS = 1;")
         self.stdout.write(self.style.SUCCESS("Existing data erased"))
 
         self.stdout.write("Migrating database...")
+        call_command("makemigrations")
         call_command("migrate")
 
         self.stdout.write("Generating Data")
@@ -64,6 +68,8 @@ class Command(BaseCommand):
         self.create_canopeum_site()
         self.create_post_conopeum_site()
         self.create_batch()
+        self.create_roles()
+        self.create_user()
         self.stdout.write(self.style.SUCCESS("Data Generated"))
 
     def create_fertilizer_types(self):
@@ -258,4 +264,19 @@ class Command(BaseCommand):
             total_number_seed=100,
             total_propagation=100,
             updated_at=timezone.now(),
+        )
+
+    def create_roles(self):
+        Role.objects.create(name="User")
+        Role.objects.create(name="Admin")
+        Role.objects.create(name="MegaAdmin")
+
+    def create_user(self):
+        User.objects.create_user(
+            username="admin",
+            email="admin@beslogic.com",
+            password="Adminbeslogic!",  # noqa: S106 MOCK_PASSWORD
+            is_staff=True,
+            is_superuser=True,
+            role=Role.objects.get(name="Admin"),
         )
